requirements.txt
streamlit
pandas
numpy
rdkit-pypi
py3Dmol
pytest
